Avoid replacing newlines with spaces early in the query
The parser split off the first three elements of the cell, to check
if they have a connection string or push the result to a variable.
But if these elements weren't found, the components were glued back
together with spaces.  If one of these components was a line comment,
it ends up commenting out more than it should.

This solution tries splitting off bits to see if they match the
expected form.  But if they don't, it goes back to using the pre-
split string, thus preserving the original whitespace.  This should
allow the connection string and redirect to appear after any number
of newline characters, as long as they are still the first non-
whitespace elements in the cell.