Enable syntax highlighting for %%sql cells
This should work with a variety of older and newer versions of Jupyter.
It tries to avoid clobbering any existing highlighting modes for SQL,
but that's probably overkill.  Restarting and rerunning the import will
end up with multiple identical regexes in the list, but that doesn't
seem to be a problem.  Fixing this would be difficult because, in JS,
/a/ !== /a/.

This is not working in JupyterLab, which seems to have a whole different
way to deal with syntax highlighting.